requirements.txt

streamlit
deepface
opencv-python-headless
spotipy
Pillow
numpy
